% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emissions.R
\docType{data}
\name{emissions}
\alias{emissions}
\title{Total emissions by economic activity}
\format{
A list of matrices:
\describe{
\item{Air pollutants}{A list of vectors containing air pollutant emissions by sector}
\item{GHGs}{Greenhouse gases emissions by sector}
\item{Heavy Metals}{Heavy metal emissions by sector}
}
}
\usage{
emissions
}
\description{
A list of lists containing air pollutant, GHGs and heavy metals emissions by economic activity in the reference year.
Data re from ISTAT.
}
\keyword{datasets}
