% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\docType{data}
\name{employment}
\alias{employment}
\title{Employment vectors}
\format{
A list of vectors:
\describe{
\item{H}{Hours worked}
\item{JB}{Number of jobs}
\item{PS}{Persons employed}
}
}
\usage{
employment
}
\description{
A list of vectors containing information on workers, hours worked and jobs for each sector of the economy
}
\keyword{datasets}
