% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous_vectors.R
\name{exogenous_vectors}
\alias{exogenous_vectors}
\title{Preparing exogenous vectors for multipliers analysis}
\usage{
exogenous_vectors(
  input_vector,
  new_sector_name = "New Sector",
  sector = "V27",
  input_value
)
}
\arguments{
\item{input_vector}{A rearranged exogenous vector.}

\item{new_sector_name}{The name of the new sector}

\item{sector}{The name of the sector from which the new sector value must be subtracted from}

\item{input_value}{The value to be added to the vector}
}
\description{
This function adds the value corresponding to a new sector to a rearranged exogenous vector (i.e. jobs vector, or emission vectors).
At the same time the value of the new sector is substracted from the (user-defined) sector of provenance.
}
