% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_oecd.R
\docType{data}
\name{legend_oecd}
\alias{legend_oecd}
\title{Legend}
\format{
A list of tibbles:
\describe{
\item{products}{A tibble contaning OECD and ISTAT product codes and labels}
\item{sectors}{A tibble contaning OECD and ISTAT sector codes and labels}
\item{order}{A tibble with the correct order of ISTAT labels in SUT tables}
\item{final_demand}{A tibble contaning OECD final demand codes and labels}
}
}
\usage{
legend_oecd
}
\description{
A list containing legends of the variables and codes used in the package
}
\keyword{datasets}
