% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_input.R
\name{loading_input}
\alias{loading_input}
\title{Loading the input file}
\usage{
loading_input(input_path, new_sector = TRUE, new_sector_name = "New Sector")
}
\arguments{
\item{input_path}{The path to the .xlsx file to be loaded}

\item{new_sector}{A logical value. If set to `TRUE` the impact assessment is carried out introducing a new sector into the economy.
If it is set to `FALSE` the simulation refer to a change in final demand and considers existing sectors only.}

\item{new_sector_name}{A string. The name of the new sector/product to be created in order to assess its impacts.}
}
\value{
A list of matrices and vectors with input data to be used to build the SUT model
}
\description{
This function loads an excel file containing input data for the SUT model. The excel file must follow a given template.
The function returns a list of matrices containing all the necessary inputs to carry out the impact analysis based on the
supply and use tables.
}
