% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tables_new_sector.R
\name{prepare_tables_new_sector}
\alias{prepare_tables_new_sector}
\title{Preparing Supply and Use tables including a new sector}
\usage{
prepare_tables_new_sector(
  input_path,
  new_sector_name = "New Sector Name",
  sector = "V27",
  transport_distribution_services = c("RD", "R45", "R46", "R47", "R49", "R50", "R51"),
  sector_to_join = "VT",
  product_to_join = "RT",
  destination_sector = "V96",
  destination_product = "R96"
)
}
\arguments{
\item{input_path}{The path to the .xlsx file to be loaded}

\item{new_sector_name}{A string. The name of the new sector/product to be created in order to assess its impacts.}

\item{sector}{The ISTAT code of the sector from which the "New" sector must be substracted from to void double counting}

\item{transport_distribution_services}{Transport and distribution services (those to which distribution and transport margins flow into,
when passing from purchasers' to base prices)}

\item{sector_to_join, product_to_join}{ISTAT codes of the sector and product to be aggregated to other sectors/products. The full list
of products/sectors is available in `inputoutputR::legend_oecd`. Default values are: `VT` and `RT` respectively.}

\item{destination_sector, destination_product}{ISTAT codes of the sector and products to which `sector_to_join` and `product_to join`
must be aggregated to. Default values are: `V96` and `R96` respectively.}
}
\value{
A list containing rearranged tables (with a new sector/product) and exogenous inputs.
}
\description{
This function takes the original S&U tables and rearranges them to make room for a new sector/product in order to be able to
perform an impact analysis.
}
