% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_input.R
\name{preparing_input}
\alias{preparing_input}
\title{Preparing input matirces}
\usage{
preparing_input(
  input_path,
  sector = "V27",
  transport_distribution_services = c("RD", "R45", "R46", "R47", "R49", "R50", "R51"),
  new_sector_name = "New Sector"
)
}
\arguments{
\item{input_path}{The path to the .xlsx file to be loaded}

\item{sector}{The ISTAT code of the sector from which the "New" sector must be substracted from to void double counting}

\item{transport_distribution_services}{Transport and distribution services (those to which distribution and transport margins flow into,
when passing from purchasers' to base prices)}

\item{new_sector_name}{A string. The name of the new sector/product to be created in order to assess its impacts.}
}
\description{
This function loads input data on the new sector/product to be assessed and computes the taxes vector and the
use and inport use vectors at base prices.
}
