% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_tables.R
\name{rearrange_tables}
\alias{rearrange_tables}
\title{Rearranging national accounting tables}
\usage{
rearrange_tables(
  sector_to_join = "VT",
  product_to_join = "RT",
  destination_sector = "V96",
  destination_product = "R96",
  type = "sut"
)
}
\arguments{
\item{sector_to_join, product_to_join}{ISTAT codes of the sector and product to be aggregated to other sectors/products. The full list
of products/sectors is available in `inputoutputR::legend_oecd`. Default values are: `VT` and `RT` respectively.}

\item{destination_sector, destination_product}{ISTAT codes of the sector and products to which `sector_to_join` and `product_to join`
must be aggregated to. Default values are: `V96` and `R96` respectively.}

\item{type}{A lstring indicating what type of tables must be rearranged. Three values possible: `sut`, `employment` and `emissions`.}
}
\value{
A list of matrices, of the same length of `inputoutputR::tables`.
}
\description{
This function takes OECD national accounting matrices and rearranges them by adding up sectors/products chosen by the user to other
sectors/products chosen by the user. The output is a list of matrices with one dimension less than the original matrices in order to
"make room" for a new sector and a new product. In this way it is possible to evaluate the socio-economic impact of the new product/sector
on the economy.
}
\examples{
# rearrange_tables()

}
