% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sut_mono.R
\name{sut_mono}
\alias{sut_mono}
\title{Supply and Use model}
\usage{
sut_mono(
  type = "type1",
  sector = "V27",
  new_sector_name = "FAAM",
  transport_distribution_services = c("RD", "R45", "R46", "R47", "R49", "R50", "R51"),
  prices = "pb",
  sector_to_join = "VT",
  product_to_join = "RT",
  destination_sector = "V96",
  destination_product = "R96"
)
}
\arguments{
\item{type}{A string that can assume two values: either `type1` or `type2`. In case `type1` is specified, final consumption is trated
as an exogenous variable (direct and indirect effect only). If `type2` is specified, final consumption is trated as endgenous and total impact include direct,
indirect and induced effects.}

\item{sector}{The ISTAT code of the sector to which the new product/sector belongs in the real world and from which the latter must
be extrapolated.}

\item{new_sector_name}{The name of the new sector/product to be created}

\item{transport_distribution_services}{A vector of ISTAT product/service codes denoting transportation and distribution services
(needed to convert inputs, expressed at purchasers' prices into base prices)}

\item{prices}{A string denoting whether the analysis must be carried out at base (`pb`) or pruchasers' (`pa`) prices. The former are
used to assess the impact of the new sector in full operation, while the latter to assess the impact of the building phase.}

\item{sector_to_join, product_to_join}{Sector and product ISTAT codes of the sector/product to be grouped with sector `V96` and product
`R26` respectively to "make" room for the new sector/products into the matrices.}

\item{destination_sector, destination_product}{Sector and product ISTAT codes of the sector/product to which `sector_to_join` and
`product_to_join` must be added to.}
}
\value{
A list with all simulation results: overall impacts on output, value added, employment, imports and multipliers.
}
\description{
This function builds and resolve a supply and use model to assess the impact generated by the creation of a new product/sector in
the Italian economy. Supply and Use tables are from the OECD. The function loads an user-defined input spreadsheet containing
information on inputs used by the new sectors, their provenance, people working in the new sector, etc. This file should respect a given
format. The function returns the overall impact on output, intermediate and final imports, value added, gdp and employment, as well as
multipliers' matrices (output, value added, jobs created).
}
\examples{
# sut_mono()
}
