% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{data}
\name{tables}
\alias{tables}
\title{SUT Tables}
\format{
A list of matrices:
\describe{
\item{supply}{Supply table at basic prices}
\item{use_pa}{Use table at purchasers' prices}
\item{use_t_pa}{Use import table at purchasers' prices}
\item{use_pb}{Use table at base prices}
\item{use_t_pb}{Use import table at base prices}
\item{margins}{Margins table at base prices}
\item{final_use_pa}{final demand PA}
\item{final_use_pb}{final demand PB}
\item{final_use_t_pa}{Final use import PA}
\item{final_use_t_pb}{Final use import PB}
\item{final_taxes}{Final taxes less subsidies}
\item{taxes}{Tax table at base prices}
\item{va}{Value added and total output}
}
}
\usage{
tables
}
\description{
A list containing SUT tables referring to Italy in matrix form. Data are from the OECD. Sectors and products
are aggregated following Istat's classification. Data refer to 2019.
}
\keyword{datasets}
